Mid-02/requirements.txt
web3==6.0.0
python-dotenv==0.19.0
google-generativeai>=0.3.0
py-solc-x>=2.0.0